refactor(health): move the canary round trip into its own service
Review feedback: the route module had grown the failure classification and
the outbound transport, which is the service layer's work — the route parses
the request and serializes the response. HealthCheckFailedError, sendCanary
and the canary deadline move to server/health-probes, tests with them.

Also from review: name the prefetch boolean isCustomerActionable, and put
the shutdown phase's when-blocks under the given-block the suite convention
asks for.
