+++ platform/app/src/server/health-probes/__tests__/canary.service.unit.test.ts
@@ -9,7 +9,7 @@ vi.mock("@langwatch/observability", () => ({
   createLogger: () => logger,
 }));
 
-import { sendCanary } from "../health-checks";
+import { sendCanary } from "../canary.service";
 
 function canary() {
   return sendCanary({
